moved export, toggle needs... WORK, fixed (i think) canvas and rubric issue, need space for upload on top
