+++ client/package.json
@@ -30,6 +30,7 @@
     "react-rnd": "^10.4.1",
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
+    "startbootstrap-simple-sidebar": "^6.0.6",
     "url-search-params": "^1.1.0",
     "web-vitals": "^2.1.4",
     "xlsx": "^0.18.5",
