+++ client/src/assets/css/UserPage.css
@@ -5,9 +5,9 @@
 */
 
 .userPage {
-  width: 95%;
+  width: 100%;
   height: auto;
-  margin: 20px 0px 0px 20px ;
+  margin: 20px 10px 40px 10px;
 }
 
 html {
@@ -380,15 +380,16 @@ Display for five
  *          PDF / CANVAS
  *
 */
-
+/*
 .docView {
-  width: 100%; 
-  height: 182.5vh;
+  width: 80%; 
+  height: 90vh;
+  max-height: 100%;
   float: left;
 }
 /*
 Display for docView
-*/
+*
 @media screen and (max-width: 1150px) {
   .docView {
     width: 200%;
@@ -396,7 +397,7 @@ Display for docView
     height: 100%;
   }
 }
-
+*/
 
 /*
 Display for docView
@@ -404,7 +405,8 @@ Display for docView
 
 
 .fileView {
-  height: 50%;
+  width: 98%;
+  height: 90vh;
   border: 1px solid #708090;
   overflow-y: auto;
   overflow-x: hidden;
@@ -433,7 +435,7 @@ Display for docView
 .react-pdf__Page__textContent,
 .react-pdf__Page__annotations, .annotationLayer,
 .textLayer {
-  display: none;
+  width: 90% !important;
 }
 
 /*
@@ -443,13 +445,17 @@ Display for docView
 */
 
 .rubricBoxContainer {
- float: right;
-  overflow-y: none;
-  border-radius: 10px;
-  width : 19vw;
-  height:45vw;
+  top: 0px;
+  right: 5px;
+  position: absolute;
+  margin-right: 14;
+  width : 20%;
+  height: 100%;
+  border-radius: 1rem;
   padding: 0;
+  z-index: 1;
 }
+
 /*
 Display for rubricBox container
 */
@@ -464,15 +470,17 @@ Display for rubricBox container
   }
 }
 
+*/
+/*
 @media screen and (min-width: 1151px) {
-  .rubricBoxContainer {
-    /* margin-top: 100px; */
-    
-  /*  width: 32vw;
+  .rubricBoxContainer {    
+    width: 32vw;
     height: 60vw;
     margin-right: -25px;
   }
 }
+*/
+/*
 @media screen and (min-width: 500px) and (max-width: 1149px) {
   .rubricBoxContainer {
     margin-top: 170px;
@@ -482,26 +490,24 @@ Display for rubricBox container
     margin-right: -0px;
   }
 }
-*/
 
+*/
 
 /*
 Display for rubricBox container k
 */
 
-/*
+
 .rubricBox {
-  border: 1px solid #708090;
   box-shadow: 0px 3px 5px -2px black;
   overflow-y: scroll;
   overflow-x: hidden;
-  height:40vw;
+  height: 70vh;
   border-radius: 10px;
   width : 100%;
   padding: 15px;
   background-color: white;
 }
-*/
 /*
 Display for rubricBox 
 */
@@ -515,16 +521,16 @@ Display for rubricBox
     margin-right: -30px;
   }
 }
-
+/*
 @media screen and (min-width: 1151px) {
-  .rubricBox {
-    /* margin-top: 100px; */
-  /*  
+  .rubricBox {    
     width: 32vw;
     height: 60vw;
     margin-right: -25px;
   }
 }
+*/
+/*
 @media screen and (min-width: 500px) and (max-width: 1149px) {
   .rubricBox {
     margin-top: 170px;
@@ -543,68 +549,6 @@ Display for rubricBox
 */
 
 
-.rubricBoxContainer{
-  position: absolute;
-  transition: right 0.4s cubic-bezier(0, 0, 0, 1);
-  overflow-x: hidden;
-  overflow-y: overlay;
-}
-
-.rubricBox {
-  border: 1px solid #708090;
-  box-shadow: 0px 3px 5px -2px black;
-  overflow-y: scroll;
-  overflow-x: hidden;
-  height:40vw;
-  border-radius: 10px;
-  width : 100%;
-  padding: 15px;
-  background-color: white;
-}
-.rubricBox li:nth-child(1){
-  padding: 0px 5px;
-  display: flex;
-  justify-content: center;
-  margin-right: 19px;
-  font-size: 14px;
-}
-
-
-.rubricBox li:hover:nth-child(1){
-  padding:10px 0px;
-  background-color:transparent;
-  color:#000;
-}
-
-.rubricBox li{
-  padding:12px;
-  cursor:pointer;
-  transition:all 0.5s;
-  letter-spacing:.5px;
-  display:flex;
-  align-items:center;
-   gap:10px;
-  border-top-left-radius:30px;
-  border-bottom-left-radius:30px;
-}
-
-.rubricBox li span:nth-child(1){
-  font-size:20px;
-}
-
-.rubricBox li span:nth-child(2){
-  font-size:16px;
-}
-
-.rubricBox li:hover{
-  background-color:#b700ff;
-  padding:12px;
-  color:#fff;
-  transition:all 0.2s;
-   
-}
-
-/*
 .rubricBox h3 {
   padding: 10px 0;
   margin: 0;
@@ -612,7 +556,6 @@ Display for rubricBox
 
 .category {
   border-radius: 10px;
-  margin: 0px;
   padding: 10px 0px;
 }
 
@@ -629,7 +572,6 @@ Display for rubricBox
 .category ul li {
   margin-left: 40px;
 }
-*/
 
 .category ul li button {
   display: inline-block;
@@ -671,6 +613,7 @@ Display for rubricBox
   list-style: none;
   display: flex;
   font-size: 18px;
+  
 }
 
 #toolbar {
@@ -695,9 +638,9 @@ Display for rubricBox
   border-radius: 5px;
   font-size: 14px;
   cursor: pointer;
-  margin: 18px 0px;
+  margin: .5% 0 5% 2%;
   text-decoration: none;
-  float: right;
+  float: left;
 }
 
 /*
@@ -707,15 +650,13 @@ Display for rubricBox
 */
 
 .critiqueBoxContainer {
-  /* border: 1px solid black; */
   width: 100%;
   height: 35%;
-  margin: 0px auto;
-  margin-top: 55px;
+  margin: 0 0 0 2rem;
 }
 .critiqueBox {
-  width: 60%;
-  margin: auto;
+  width: 70%;
+  margin: 0px;
 }
 
 .critiqueBox h4{
@@ -752,20 +693,10 @@ Display for rubricBox
 }
 
 .commentBoxButtons {
-  float: right;
-  margin: 0 13%;
+  margin: 0 3%;
 }
 
-.submitButton {
-  background-color: #7C4DFF;
-  color: white;
-  padding: 6px 20px;
-  border: none;
-  border-radius: 5px;
-  font-size: 16px;
-  cursor: pointer;
-  margin: 10px 10px 20px 20px;
-}
+
 
 /*
  *
@@ -846,7 +777,7 @@ Display for rubricBox
 */
 
 .pageNavigation {
-  float: right;
+  float: left;
   margin: 10px 0px;
 }
 
@@ -935,6 +866,5 @@ input[type="file"] {
   background-color: #9f82ec;
   color: #fff;
 }
-
