+++ client/src/userpage/CritiqueBox.js
@@ -1,7 +1,6 @@
 import React, {useContext, useState, useEffect} from 'react';
 import { ItemContext } from './ItemContext';
 import Item from './Item';
-import Draggable from "react-draggable";
 
 
 export default function CritiqueBox(){
@@ -36,9 +35,8 @@ export default function CritiqueBox(){
         return (
             <div className='critiqueBoxContainer' >
                 <div className="critiqueBox">
-                    <h3 style={{margin: '10px 0'}}>{item1.path}</h3>
-                    <div style={{display: 'flex', alignItems: 'center', height:'40px'}}>
-                        <h4 style={{marginRight: '5px', marginTop: '0px'}}>Rating: </h4>
+                    <div style={{margin: '0', width: '15vw'}}>
+                        <h4> Rating: </h4>
                         <div style={{display: 'flex'}}>
                             <button style={{visibility: `${item1.Comment[0] === '' ? 'visible' : 'hidden'}` }} id="one" onClick={() => {setIndex1(0)}}>1</button>
                             <button style={{visibility: `${item1.Comment[1] === '' ? 'visible' : 'hidden'}` }} id="two" onClick={() => {setIndex1(1)}}>2</button>
@@ -50,8 +48,9 @@ export default function CritiqueBox(){
                     </div>
                     <textarea id="commentArea" type="text" value={comment} onChange={handleCommentChange} style={{margin: '10px 0'}}/>
                     <span style={{ float: 'right', margin: '5px 0' }}>
-                        <button className="generalButton"  onClick={() => { saveReturn(comment) }}>Submit</button>
+                        <button className="generalButton" onClick={() => { saveReturn(comment) }}>Submit</button>
                     </span>
+                    <br/>
                 </div>
             </div>
