+++ client/src/userpage/RubricBox.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState, useContext, useRef } from 'react'
 import { ItemContext } from "./ItemContext"
 import { CSVLink } from "react-csv";
+import CritiqueBox from './CritiqueBox';
 
 export default function RubricBox() {
   let {totalItems, Hierarchy} = useContext(ItemContext);
@@ -16,13 +17,6 @@ export default function RubricBox() {
       setdirjsonExists(data);
     });
   }, []);
-
-  function handleExport(event) {
-    const confirmed = window.confirm('Are you sure you want to export the results?');
-    if (!confirmed) {
-      event.preventDefault(); // prevent the default behavior of the onClick event
-    }
-  }
 
   if (dirjsonExists) {
     return (
@@ -45,19 +39,12 @@ export default function RubricBox() {
           {isOpen || isTransitioning ? (
             <>
               <h3>Available Categories</h3>
-              {Hierarchy.map((item, i) => <div key={i}>{item.returnHTML()}</div>)}
+              {Hierarchy.map((item, i) => <div key={i}>{item.returnHTML()} <CritiqueBox/> </div>
+              )}
             </>
           ) : null}
         </div>
-        {/* Export button */}
-        <CSVLink 
-          data={totalItems}
-          filename={"Export_Results.csv"}
-          className='csvLink'
-          onClick={handleExport}
-        >
-          Export
-        </CSVLink>
+
       </div>
     );
   } else {
